% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_alexnet_mod.R
\name{alexnet_mod}
\alias{alexnet_mod}
\title{AlexNet}
\usage{
alexnet_mod(pretrained = FALSE, progress = TRUE, ...)
}
\description{
AlexNet Model Architecture (copy from https://github.com/mlverse/torchvision/blob/main/R/models-alexnet.R)
}
