% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrenamiento_imagenes.R
\name{entrenamiento_imagenes}
\alias{entrenamiento_imagenes}
\title{Entrenamiento de imagenes}
\usage{
entrenamiento_imagenes(
  image_path,
  model = "alexnet",
  pretrained = F,
  batch_size = 128,
  epochs = 10,
  learning_rate = 0.01,
  optimizer = "adam"
)
}
\description{
Entrenamiento de imagenes
}
