% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_loading.R
\name{image_loading}
\alias{image_loading}
\title{Image Loading}
\usage{
image_loading(image_directory)
}
\description{
Una función para cargar una base de datos compuesta por imágenes. La base de datos deberá estar compuesta por tres carpetas (train, valid y test)
}
