% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_train.R
\name{image_train}
\alias{image_train}
\title{Entrenamiento de imágenes}
\usage{
image_train(
  image_path,
  model = "resnet34",
  pretrained = FALSE,
  batch_size = 128
)
}
\description{
Entrenamiento de imágenes
}
