% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_train_lr.R
\name{image_train_lr}
\alias{image_train_lr}
\title{image_train_learning_rate}
\usage{
image_train_lr(
  image_path,
  model = "resnet34",
  pretrained = FALSE,
  batch_size = 128
)
}
\description{
image_train_learning_rate
}
\author{
Egoitz Carral
}
