% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_train_pruebas.R
\name{image_train_pruebas}
\alias{image_train_pruebas}
\title{Entrenamiento de imágenes (pruebas)}
\usage{
image_train_pruebas(
  model = "alexnet",
  pretrained = FALSE,
  batch_size = 32,
  datasets = "dogs_cats",
  image_path
)
}
\description{
Entrenamiento de imágenes (pruebas)
}
