% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prediction}
\alias{prediction}
\title{Prediction}
\usage{
prediction(model_path, object_path)
}
\arguments{
\item{model_path}{Path to the trained model}

\item{object_path}{Path to the object that is going to be classified}
}
\description{
Prediction
}
